backend/pyproject.toml
[project]
name = "cornucopia"
version = "0.1.0"
description = ""
authors = [
    {name = "WJ-Tien",email = "chc61394@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.3.2,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
